+++ include/speak_easy_2.h
@@ -20,7 +20,8 @@ typedef struct {
   bool verbose; // Print information to stdout
 } se2_options;
 
-int speak_easy_2(igraph_t* graph, igraph_vector_t* weights,
-                 se2_options* opts, igraph_vector_int_t* res);
-
+igraph_error_t speak_easy_2(igraph_t* graph, igraph_vector_t* weights,
+                            se2_options* opts, igraph_vector_int_t* res);
+igraph_error_t se2_order_nodes(igraph_vector_int_t const* memb,
+                               igraph_vector_int_t* ordering);
 #endif